+++ Source/Examples/Xamarin.Forms/IssueDemos/IssueDemos.Droid/MainActivity.cs
@@ -17,7 +17,7 @@ namespace IssueDemos.Droid
             base.OnCreate(bundle);
 
             global::Xamarin.Forms.Forms.Init(this, bundle);
-            OxyPlot.Xamarin.Forms.Platform.Android.Forms.Init();
+            OxyPlot.Xamarin.Forms.Platform.Android.PlotViewRenderer.Init();
             LoadApplication(new App());
         }
     }