+++ Source/Examples/Xamarin.Forms/SimpleDemo2/SimpleDemo2.iOS/AppDelegate.cs
@@ -23,7 +23,7 @@ namespace SimpleDemo2.iOS
 		public override bool FinishedLaunching(UIApplication app, NSDictionary options)
 		{
 			global::Xamarin.Forms.Forms.Init ();
-            OxyPlot.Xamarin.Forms.Platform.iOS.Forms.Init ();
+            OxyPlot.Xamarin.Forms.Platform.iOS.PlotViewRenderer.Init ();
 			LoadApplication (new SimpleDemo2.App ());
 
 			return base.FinishedLaunching (app, options);